test(delegation): consolidate doc upload parity tests into the delegation-auth test file
Fold the doc upload delegation-parity tests into the existing
doc_delegation_auth_test.go (where the other delegation tests live) and
remove the standalone doc_upload_delegation_parity_test.go, so delegation
tests stay in one discoverable place instead of a per-change file.
No behavior change; test names and assertions are unchanged.
